Refactor: Extract helpers and foundation slices from monolithic gameStore
Phase 1 of Zustand refactoring - extract helper functions and foundation slices
with zero dependencies to reduce gameStore.ts size.

**Helpers Extracted:**
- gulagHelpers.ts (3 functions)
- wealthCalculation.ts (calculateTotalWealth)
- playerStats.ts (initializePlayerStats)
- eliminationHelpers.ts (getEliminationMessage)

**Slices Extracted:**
- uiSlice: UI state (pendingAction, isRolling, pendingBribes) + bribe actions
- logSlice: Game log (gameLog) + addLogEntry action
- statisticsSlice: Game statistics + updatePlayerStat, calculateFinalStats actions

**New Type System:**
- storeTypes.ts: Centralised type definitions for GameStore, GameActions, and slices
- Resolved circular dependencies using proper import hierarchy

**Quality Gates:**
- ✅ All 913 tests passing
- ✅ Zero lint errors
- ✅ Coverage: 94.55% (>85% threshold)
- ✅ gameStore.ts reduced by ~200 lines
